requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
